v0.8 WS1: mypy clean at zero errors, enforced in CI
- Widen MetricsDict to include int/bool values
- Narrow Optional design-variable bounds with asserts after validation
- Cast scipy/numpy Any returns to float; annotate heterogeneous
  return dicts as dict[str, Any]
- Fix matplotlib typing (Figure cast, get_cmap, 3D axes cast)
- Add types-PyYAML dev dependency and blocking mypy step to CI (3.12)

+++ src/phased_array_systems/models/antenna/taper.py
@@ -104,7 +104,7 @@ def compute_taper_efficiency(taper: NDArray[np.floating]) -> float:
     if sum_weights_sq == 0:
         return 0.0
 
-    return (sum_weights**2) / (n * sum_weights_sq)
+    return float((sum_weights**2) / (n * sum_weights_sq))
 
 
 def taper_loss_from_sll(
@@ -239,7 +239,7 @@ def beamformer_noise_factor(
     if t_uniform == 0:
         return 1.0
 
-    return t_eff / t_uniform * n / (sum_weights**2 / np.sum(taper**2))
+    return float(t_eff / t_uniform * n / (sum_weights**2 / np.sum(taper**2)))
 
 
 def estimate_taper_parameters(
